feat(ClusterDropdown): authenticated cluster header label
Commit adds authenticated cluster dropdown label mechanism
to display the correct info when metadata available to user
or to display the hard coded string.

Closes DCOS-39679
