+++ src/js/components/HeaderBar.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import { Link } from "react-router";
 import { HeaderBar as UIHeaderBar } from "ui-kit-stage/HeaderBar";
-import ClusterDropdown from "./ClusterDropdown";
+import ClusterHeader from "./ClusterHeader";
 import AccountDropdown from "./AccountDropdown";
 import SidebarToggle from "./SidebarToggle";
 
@@ -20,7 +20,7 @@ export default function HeaderBar() {
         <span className="header-bar-logo" />
       </Link>
       <AccountDropdown />
-      <ClusterDropdown />
+      <ClusterHeader />
     </UIHeaderBar>
   );
 }
